# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
  httpErrors.RestErr:
    type: object
...
  model.Portfolio:
    properties:
      active:
        type: boolean
      created_at:
        type: string
      currency_id:
        type: string
      portfolio_id:
        type: string
      title:
        type: string
      updated_at:
        type: string
      user_id:
        type: string
    required:
    - active
    - currency_id
    - title
    - user_id
    type: object
  model.PortfolioList:
    properties:
      asset_total:
        items:
          $ref: '#/definitions/models.AssetTotal'
        type: array
      portfolio:
        $ref: '#/definitions/model.Portfolio'
    type: object
...
  models.AssetTotal:
    properties:
      averagePurchasePrice:
        type: number
      firstNotationAt:
        type: string
      identify:
        type: string
      image_url:
        type: string
      marketId:
        type: string
      ticker:
        type: string
      title:
        type: string
      totalAmount:
        type: number
      totalCount:
        type: integer
      totalQuantity:
        type: integer
    type: object
  models.Market:
    properties:
      content:
        type: string
      created_at:
        type: string
      currency_id:
        type: string
      image_url:
        type: string
      instrument_id:
        type: string
      market_id:
        type: string
      ticker:
        type: string
      title:
        type: string
      updated_at:
        type: string
    required:
    - currency_id
    - instrument_id
    - ticker
    - title
    type: object
  models.MarketBase:
    properties:
      content:
        type: string
      desc_instrument:
        type: string
      image_url:
        type: string
      market_id:
        type: string
      ticker:
        type: string
      title:
        type: string
      title_instrument:
        type: string
      updated_at:
        type: string
    required:
    - desc_instrument
    - ticker
    - title
    - title_instrument
    type: object
  models.MarketList:
    properties:
      has_more:
        type: boolean
      markets:
        items:
          $ref: '#/definitions/models.MarketBase'
        type: array
      page:
        type: integer
      size:
        type: integer
      total_count:
        type: integer
      total_pages:
        type: integer
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: REST API
  title: REST API
  version: "1.0"
paths:
...
  /market/{id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: market_id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            type: string
      security:
      - Auth: []
      summary: Удалить ценную бумагу/товар
      tags:
      - Market
    get:
      consumes:
      - application/json
      parameters:
      - description: market_id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Market'
      security:
      - Auth: []
      summary: Получить по id ценную бумагу/товар
      tags:
      - Market
    put:
      consumes:
      - application/json
      parameters:
      - description: market_id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Market'
      security:
      - Auth: []
      summary: Изменить данные ценной бумаги/товара
      tags:
      - Market
  /market/all:
    get:
      consumes:
      - application/json
      description: Получить все ценные бумаги/товары с разбивкой на страницы
      parameters:
      - description: page number
        format: page
        in: query
        name: page
        type: integer
      - description: number of elements per page
        format: size
        in: query
        name: size
        type: integer
      - description: filter name
        format: orderBy
        in: query
        name: orderBy
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.MarketList'
      security:
      - Auth: []
      summary: Получить все ценные бумаги/товары
      tags:
      - Market
  /market/create:
    post:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Market'
      security:
      - Auth: []
      summary: Добавить новую ценную бумаги/товара
      tags:
      - Market
  /market/parse:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: ""
      security:
      - Auth: []
      summary: Спарсить
      tags:
      - Market
  /market/search:
    get:
      consumes:
      - application/json
      description: Искать ценную бумагу/товар по названию
      parameters:
      - description: title
        in: query
        name: title
        type: integer
      - description: page number
        format: page
        in: query
        name: page
        type: integer
      - description: number of elements per page
        format: size
        in: query
        name: size
        type: integer
      - description: filter name
        format: orderBy
        in: query
        name: orderBy
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.MarketList'
      security:
      - Auth: []
      summary: Поиск по названию
      tags:
      - Market
  /portfolio/active-total:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.PortfolioList'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpErrors.RestErr'
      security:
      - Auth: []
      summary: Портфель пользователя
      tags:
      - Portfolio
securityDefinitions:
  Auth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"